s3-checksum256/pyproject.toml
[tool.poetry]
name = "s3-checksum256"
version = "0.1.0"
description = "Container for calculating SHA 256 checksum on objectstore objects."
authors = ["Per Buer <per.buer@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.7"
py-objectstore = { git = "https://github.com/arkivverket/py-objectstore.git" }

[tool.poetry.dev-dependencies]
pylint = "^2.4.4"
python-dotenv = "^0.13.0"
autopep8 = "^1.5.3"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
